+++ inst/shiny-examples/bucket_list/app.R
@@ -4,7 +4,6 @@
 library(shiny)
 library(sortable)
 
-
 ui <- fluidPage(
   tags$head(
     tags$style(HTML(".bucket-list-container {min-height: 350px;}"))
@@ -12,8 +11,6 @@ ui <- fluidPage(
   fluidRow(
     column(
       tags$b("Exercise"),
-      actionButton("btnUpdateBucket", label = "Update bucket list title"),
-      actionButton("btnUpdateRank", label = "Update rank list title"),
       width = 12,
       bucket_list(
         header = "Drag the items in any desired bucket",
@@ -76,29 +73,6 @@ server <- function(input, output, session) {
       input$bucket_list_group # Matches the group_name of the bucket list
     )
 
-  # test updating the bucket list label
-  counter_bucket <- reactiveVal(1)
-  observe({
-    update_bucket_list(
-      "bucket_list_group",
-      text = paste("You pressed the button", counter_bucket(), "times"),
-      session = session
-    )
-    counter_bucket(counter_bucket() + 1)
-  }) %>%
-    bindEvent(input$btnUpdateBucket)
-
-  # test updating the rank list label
-  counter_rank <- reactiveVal(1)
-  observe({
-    update_rank_list(
-      "rank_list_1",
-      text = paste("You pressed the button", counter_rank(), "times"),
-      session = session
-    )
-    counter_rank(counter_rank() + 1)
-  }) %>%
-    bindEvent(input$btnUpdateRank)
 }