+++ inst/shiny-examples/rank_list/app.R
@@ -48,7 +48,6 @@ ui <- fluidPage(
         tabPanel(
           "Default",
             tags$b("Exercise"),
-            actionButton("btnUpdate", label = "Update rank list title"),
             rank_list_basic,
             tags$b("Result"),
             verbatimTextOutput("results_basic")
@@ -82,15 +81,6 @@ server <- function(input, output, session) {
   output$results_swap <- renderPrint({
     input$rank_list_swap # This matches the input_id of the rank list
   })
-  # test updating the rank list label
-  observe({
-    update_rank_list(
-      "rank_list_basic",
-      text = paste("You pressed the button at", as.character(Sys.time())),
-      session = session
-    )
-  }) %>%
-    bindEvent(input$btnUpdate)
 }
 
 shinyApp(ui, server)