+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ## Enhancements
 
-* Add experimental support for `update_rank_list()` and `update_bucket_list()`
+* Add support for `update_rank_list()` and `update_bucket_list()`
 * Add ability to switch the orientation of `rank_list()` items to horizontal. #92
 
 ## Changes